# internal/app/docs/swagger.yaml
basePath: /
definitions:
...
  ds.DeleteFlightReq:
    properties:
      id:
        type: integer
    type: object
  ds.DeleteFlightRes:
    properties:
      flight_id:
        type: integer
      status:
        type: string
    type: object
...
  ds.FlightRequest:
    properties:
      ams:
        type: string
      date_completion;default:'0001-01-01':
        type: string
      date_create:
        type: string
      date_formation:
        type: string
      id:
        type: integer
      moder_id:
        type: integer
      planets_request:
        items:
          $ref: '#/definitions/ds.PlanetsRequest'
        type: array
      status:
        type: string
      user_id:
        type: integer
      user_login:
        type: string
    type: object
...
  ds.FlightsListRes2:
    properties:
      flight:
        items:
          $ref: '#/definitions/ds.FlightRequest'
        type: array
      status:
        type: string
    type: object
...
  ds.Planet:
    properties:
      description:
        type: string
      distance:
        type: number
      gravity:
        type: number
      id:
        description: gorm.Model
        type: integer
      image:
        type: string
      is_delete:
        type: boolean
      name:
        type: string
      radius:
        type: number
      type:
        type: string
    type: object
...
  ds.PlanetsRequest:
    properties:
      flight_number:
        type: integer
      fr_id:
        type: integer
      planet:
        $ref: '#/definitions/ds.Planet'
      planet_id:
        type: integer
    type: object
...
  ds.UpdateFlightReq:
    properties:
      ams:
        type: string
      description:
        type: string
      id:
        type: integer
    type: object
...
  ds.UpdatedFlightRes:
    properties:
      ams:
        type: string
      date_completion:
        type: string
      date_create:
        type: string
      date_formation:
        type: string
      id:
        type: integer
      moder_id:
        type: integer
      status:
        type: string
      user_id:
        type: integer
      user_login:
        type: string
    type: object
  handler.errorResp:
    properties:
      description:
        example: Описание ошибки
        type: string
      status:
        example: error
        type: string
    type: object
host: localhost:8888
info:
  contact:
    email: drakenchef@gmail.com
    name: API Support
    url: https://github.com/Drakenchef
  description: AMS flights
  title: AMS
  version: "1.0"
paths:
  /Flights:
    delete:
      consumes:
      - application/json
      description: Удаление заявки по идентификатору.
      parameters:
      - description: Идентификатор заявки для удаления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.DeleteFlightReq'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное удаление заявки
          schema:
            $ref: '#/definitions/ds.DeleteFlightRes'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResp'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResp'
      security:
      - ApiKeyAuth: []
      summary: Удаление заявки
      tags:
      - Заявки
    get:
      description: Получение списка заявок с фильтрами по статусу, дате начала и дате
        окончания, пользователю.
      parameters:
      - description: Статус заявки.
        in: query
        name: status
        type: string
      - description: Дата начала периода фильтрации в формате '2006-01-02'.
        in: query
        name: date_formation_start
        type: string
      - description: Дата окончания периода фильтрации в формате '2006-01-02'.
        in: query
        name: date_formation_end
        type: string
      - description: Сортировка по login пользователя.
        in: query
        name: user_login
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список заявок
          schema:
            items:
              $ref: '#/definitions/ds.FlightsListRes2'
            type: array
        "204":
          description: Нет данных
          schema:
            $ref: '#/definitions/handler.errorResp'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResp'
      security:
      - ApiKeyAuth: []
      summary: Список заявок
      tags:
      - Заявки
    put:
      consumes:
      - application/json
      description: Обновление данных о заявке.
      parameters:
      - description: Данные для обновления заявки
        in: body
        name: updatedFLight
        required: true
        schema:
          $ref: '#/definitions/ds.UpdateFlightReq'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление данных о заявке
          schema:
            $ref: '#/definitions/ds.UpdatedFlightRes'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResp'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResp'
      security:
      - ApiKeyAuth: []
      summary: Обновление данных о заявке
      tags:
      - Заявки
  /Flights/{id}:
    get:
      description: Получение информации о заявке по его ID.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Информация о заявке по ID
          schema:
            $ref: '#/definitions/ds.FlightsListRes2'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResp'
        "404":
          description: Заявка не найдена
          schema:
            $ref: '#/definitions/handler.errorResp'
      summary: Получение информации о заявке по её ID.
      tags:
      - Заявки
  /FlightsModer/{id}:
    put:
      consumes:
      - application/json
      description: Обновление статуса заявки для модератора.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление статуса
          schema:
            type: string
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResp'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResp'
      security:
      - ApiKeyAuth: []
      summary: Обновление статуса заявки для модератора
      tags:
      - Заявки
  /FlightsUser/{id}:
    put:
      consumes:
      - application/json
      description: Обновление статуса заявки для пользователя.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление статуса
          schema:
            type: string
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResp'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResp'
      security:
      - ApiKeyAuth: []
      summary: Обновление статуса заявки для пользователя.
      tags:
      - Заявки
...
  /UsersFlight:
    get:
      description: Получение списка заявок пользователем.
      produces:
      - application/json
      responses:
        "200":
          description: Список заявок
          schema:
            items:
              $ref: '#/definitions/ds.FlightsListRes2'
            type: array
        "204":
          description: Нет данных
          schema:
            $ref: '#/definitions/handler.errorResp'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResp'
      security:
      - ApiKeyAuth: []
      summary: Список заявок пользователя
      tags:
      - Заявки
  /UsersFlightUpdate:
    put:
      consumes:
      - application/json
      description: Обновление данных о заявке пользователем.
      parameters:
      - description: Данные для обновления заявки пользователем
        in: body
        name: updatedHike
        required: true
        schema:
          $ref: '#/definitions/ds.UpdateFlightReq'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление данных о заявке пользователя
          schema:
            $ref: '#/definitions/ds.UpdatedFlightRes'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResp'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResp'
      security:
      - ApiKeyAuth: []
      summary: Обновление данных о заявке пользователем
      tags:
      - Заявки
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"